requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.2
django-crispy-forms==2.3
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2